fix an issue with signature generation where modeling statements would have the auto-generated internal "at" position member count against whether to have the first parameter listed as anonymous ex. should be "TestStatement:test:" not "TestStatement testName:test:"

+++ cmd/makeops/templates/sigMapTemplate.js
@@ -1,7 +1,7 @@
 // sigMap.js
 'use strict';
 module.exports =`
-var Signatures = map[uint]interface{}{
+var Signatures = map[uint64]interface{}{
 {{#each list}}
   {{@key}}: (*{{Pascal type}})(nil), /* {{sig}} */
 {{/each}}
